Moving matrix client info related component to core package
