+++ vector/src/main/java/im/vector/app/core/session/clientinfo/MatrixClientInfoContent.kt
@@ -14,7 +14,20 @@
  * limitations under the License.
  */
 
-/**
- * Prefix for th key account data event which holds client info.
- */
-const val MATRIX_CLIENT_INFO_KEY_PREFIX = "io.element.matrix_client_information."
+package im.vector.app.core.session.clientinfo
+
+import com.squareup.moshi.Json
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+data class MatrixClientInfoContent(
+        // app name
+        @Json(name = "name")
+        val name: String? = null,
+        // app version
+        @Json(name = "version")
+        val version: String? = null,
+        // app url (optional, applicable only for web)
+        @Json(name = "url")
+        val url: String? = null,
+)
