+++ vector/src/test/java/im/vector/app/core/session/clientinfo/GetMatrixClientInfoUseCaseTest.kt
@@ -14,9 +14,8 @@
  * limitations under the License.
  */
 
-package im.vector.app.features.settings.devices.v2.details.extended
+package im.vector.app.core.session.clientinfo
 
-import MATRIX_CLIENT_INFO_KEY_PREFIX
 import im.vector.app.test.fakes.FakeActiveSessionHolder
 import org.amshove.kluent.shouldBe
 import org.amshove.kluent.shouldBeEqualTo
